update tests for the relocation of apply_miri_scattering to detectors.py
